+++ src/NzbDrone.Core/MediaFiles/BookFileMediaType.cs
@@ -1,0 +1,9 @@
+namespace NzbDrone.Core.MediaFiles
+{
+    public enum BookFileMediaType
+    {
+        Unknown = 0,
+        Ebook = 1,
+        Audiobook = 2
+    }
+}