+++ src/NzbDrone.Core/MediaFiles/BookImport/Identification/IdentificationService.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using NLog;
 using NzbDrone.Common.Extensions;
 using NzbDrone.Common.Instrumentation.Extensions;
+using NzbDrone.Core.MediaFiles;
 using NzbDrone.Core.MediaFiles.BookImport.Aggregation;
 using NzbDrone.Core.Parser.Model;
 
@@ -114,7 +115,10 @@ namespace NzbDrone.Core.MediaFiles.BookImport.Identification
                 FileTrackInfo = _metadataTagService.ReadTags((FileInfoBase)new FileInfo(x.Path)),
                 ExistingFile = true,
                 AdditionalFile = true,
-                Quality = x.Quality
+                Quality = x.Quality,
+                MediaType = x.MediaType != BookFileMediaType.Unknown
+                    ? x.MediaType
+                    : MediaFileExtensions.GetMediaTypeForExtension(Path.GetExtension(x.Path))
             }))
             .ToList();